added script and problem structure for easily trying out problem solutions
