+++ README.md
@@ -6,11 +6,41 @@ A collaborative attempt at the problem tree found [here](http://rosalind.info/pr
 ## Repository Structure
 
 - README.md
+- scripts/
+    - run (runs a given problem solution on a given input file)
+    - test (runs the entire test suite)
 - src/
-    * problem_id/
-        - tests
+    - problem_id/
+        - tests.py
+        - util.py
         - main.py
-        - other_files.py
+            - Must export a function named `main` with signature
+
+            ```
+            (input_file_contents, *args) -> output_file_contents
+            ```
+
+            (for the cli to find).
+
+
+## Running the Test Suite
+
+To run the test suite from the project root, use
+
+```shell
+./scripts/test
+```
+
+
+## Trying out a Problem Solution
+
+You can run problem solutions on sample data from a file using `scripts/run`.  For example, from the project root, running
+
+```shell
+./scripts/run fibo ~/Downloads/rosalind_fibo.txt
+```
+
+will run the solution for problem "fibo" with the input data from `~/Downloads/rosalind_fibo.txt`.
 
 
 ## Contributing
