+++ src/fibo/main.py
@@ -1,15 +1,8 @@
-def fibo(n):
-    '''
-    Returns the nth Fibonacci number
-    '''
-    # ensure `n` is an integer
-    if not isinstance(n, int):
-        raise ValueError('argument must be an integer')
-    # ensure `n` is non-negative
-    elif n < 0:
-        raise ValueError('argument must be positive')
+# local imports
+from .util import fibo
 
-    if n == 0 or n == 1:
-        return n
-    else:
-        return fibo(n - 1) + fibo(n - 2)
+
+# export main function for cli
+def main(s):
+    # turn input into integer, then pass to `fibo`
+    return fibo(int(s))
